feat: Add barcode scanning functionality to home page
This commit adds barcode scanning functionality to the home page of the app. It replaces the previous overlay page with a new home page that includes a button to scan barcodes. When the button is pressed, the user is taken to a barcode scanner screen where they can scan barcodes. The scanned barcode is then displayed on the home page.