feat: Add hide gallery button option to barcode scanner
This commit adds a new option to the barcode scanner that allows hiding the gallery button. The `hideGalleryButton` parameter has been added to the `AiBarcodeScanner` widget, and when set to `true`, the gallery button will not be displayed on the scanner screen. This provides more flexibility for customizing the scanner UI.